+++ CHANGES.md
@@ -10,6 +10,7 @@
 - MongoDB Full: Improved initialization of transformation subsystem
 - MongoDB Adapter: Improved performance of when computing collection cardinality
   by using `collection.estimated_document_count()`
+- MongoDB Full: Optionally use `limit` parameter as number of total records
 
 ## 2024/09/19 v0.0.24
 - MongoDB Full: Refactor transformation subsystem to `commons-codec`